+++ src/store/actions/index.js
@@ -1,0 +1,17 @@
+export {
+    authentication,
+    logout,
+    checkAuthStatus
+} from './auth';
+
+export {
+    fetchTodoItems
+} from './history';
+
+export {
+    addTodoItem,
+    removeTodoItem,
+    completeTodoItem,
+    clearCompletedTodoItems,
+    clearTodayTodoItems
+} from './today';
